Remove timelord requirement for tests.
Inject pre-generated signage points and blocks instead.
