+++ tests/conftest.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 import asyncio
+from contextlib import AsyncExitStack
 import dataclasses
 import datetime
 import functools
@@ -36,6 +37,7 @@ from chia.server.start_service import Service
 from chia.simulator.full_node_simulator import FullNodeSimulator
 from chia.simulator.setup_nodes import (
     SimulatorsAndWallets,
+    setup_full_node,
     setup_full_system,
     setup_n_nodes,
     setup_simulators_and_wallets,
@@ -1120,59 +1122,65 @@ def populated_temp_file_keyring_fixture() -> Iterator[TempKeyring]:
 
 
 @pytest.fixture(scope="function")
-async def farmer_one_harvester_zero_bits_plot_filter(
+async def farmer_harvester_2_simulators_zero_bits_plot_filter(
     tmp_path: Path, get_temp_keyring: Keychain
 ) -> AsyncIterator[
     Tuple[
-        Service[Harvester, HarvesterAPI],
         Service[Farmer, FarmerAPI],
-        Service[FullNode, FullNodeSimulator],
-        BlockTools,
-    ]
-]:
-    zero_bit_plot_filter_consts = dataclasses.replace(test_constants_modified, NUMBER_ZERO_BITS_PLOT_FILTER=0)
-
-    bt = await create_block_tools_async(zero_bit_plot_filter_consts, keychain=get_temp_keyring)
-
-    async with setup_simulators_and_wallets_service(1, 0, bt.constants) as ([node], _, bt):
-        async with setup_farmer_multi_harvester(bt, 1, tmp_path, bt.constants, start_services=True) as (
-            [harvester_service],
-            farmer_service,
-            _,
-        ):
-            yield harvester_service, farmer_service, node, bt
-
-
-@pytest.fixture(scope="function")
-async def farmer_harvester_full_node_timelord_zero_bits_plot_filter(
-    tmp_path: Path, get_temp_keyring: Keychain
-) -> AsyncIterator[
-    Tuple[
         Service[Harvester, HarvesterAPI],
-        Service[Farmer, FarmerAPI],
         Union[Service[FullNode, FullNodeAPI], Service[FullNode, FullNodeSimulator]],
-        Service[Timelord, TimelordAPI],
+        Union[Service[FullNode, FullNodeAPI], Service[FullNode, FullNodeSimulator]],
         BlockTools,
     ]
 ]:
     zero_bit_plot_filter_consts = dataclasses.replace(
         test_constants_modified,
         NUMBER_ZERO_BITS_PLOT_FILTER=0,
-        NUM_SPS_SUB_SLOT=uint32(4),
+        NUM_SPS_SUB_SLOT=uint32(8),
     )
 
-    bt = await create_block_tools_async(
-        zero_bit_plot_filter_consts,
-        keychain=get_temp_keyring,
-        num_og_plots=0,
-        num_pool_plots=20,
-        num_non_keychain_plots=0,
-    )
+    async with AsyncExitStack() as async_exit_stack:
+        bt = await create_block_tools_async(
+            zero_bit_plot_filter_consts,
+            keychain=get_temp_keyring,
+            num_og_plots=0,
+            num_pool_plots=20,
+            num_non_keychain_plots=0,
+        )
 
-    async with setup_full_system(bt.constants, bt, bt, bt) as full_system:
-        async with setup_farmer_multi_harvester(bt, 1, tmp_path, bt.constants, start_services=True) as (
-            [harvester_service],
-            farmer_service,
-            _,
-        ):
-            yield harvester_service, farmer_service, full_system.node_1, full_system.timelord, bt
+        config_overrides: Dict[str, int] = {
+            "full_node.max_sync_wait": 0
+        }
+
+        bts = [
+            await create_block_tools_async(
+                zero_bit_plot_filter_consts,
+                keychain=get_temp_keyring,
+                num_og_plots=0,
+                num_pool_plots=0,
+                num_non_keychain_plots=0,
+                config_overrides=config_overrides,
+            )
+            for _ in range(2)
+        ]
+
+        simulators: List[Service[FullNode, FullNodeSimulator]] = [
+            await async_exit_stack.enter_async_context(
+                # Passing simulator=True gets us this type guaranteed
+                setup_full_node(  # type: ignore[arg-type]
+                    consensus_constants=bts[index].constants,
+                    db_name=f"blockchain_test_{index}_sim.db",
+                    self_hostname=bts[index].config["self_hostname"],
+                    local_bt=bts[index],
+                    simulator=True,
+                    db_version=2,
+                )
+            )
+            for index in range(len(bts))
+        ]
+
+        [harvester_service], farmer_service, _ = await async_exit_stack.enter_async_context(
+            setup_farmer_multi_harvester(bt, 1, tmp_path, bt.constants, start_services=True)
+        )
+            
+        yield farmer_service, harvester_service, simulators[0], simulators[1], bt
