--- tests/farmer_harvester/third_party_harvesters/config.py
@@ -1,4 +1,0 @@
-from __future__ import annotations
-
-checkout_blocks_and_plots = True
-install_timelord = True
